move sticky to mixin
