+++ src/cdk/table/can-stick.ts
@@ -6,6 +6,8 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
+import {coerceBooleanProperty} from '@angular/cdk/coercion';
+
 /** @docs-private */
 export type Constructor<T> = new(...args: any[]) => T;
 
@@ -16,6 +18,9 @@ export type Constructor<T> = new(...args: any[]) => T;
  * @docs-private
  */
 export interface CanStick {
+  /** Whether sticky positioning should be applied. */
+  sticky: boolean;
+
   /** Whether the sticky input has changed since it was last checked. */
   _hasStickyChanged: boolean;
 
@@ -34,6 +39,15 @@ export interface CanStick {
 export function mixinHasStickyInput<T extends Constructor<{}>>(base: T):
     Constructor<CanStick> & T {
   return class extends base {
+    /** Whether sticky positioning should be applied. */
+    get sticky(): boolean { return this._sticky; }
+    set sticky(v: boolean) {
+      const prevValue = this._sticky;
+      this._sticky = coerceBooleanProperty(v);
+      this._hasStickyChanged = prevValue !== this._sticky;
+    }
+    _sticky: boolean = false;
+
     /** Whether the sticky input has changed since it was last checked. */
     _hasStickyChanged: boolean = false;
